+++ resources/js/website/components/CompanyDashboardMyBillingHistory.vue
@@ -1,7 +1,7 @@
 <template>
   <div style="min-height:100px">
     <h3>{{'dashboard.billing.title'|trans}}</h3>
-    <b-overlay variant="dark" :show="!isLoaded " blur=""></b-overlay>
+    <overlay :isLoaded="isLoaded"></overlay>
     <div style="overflow:auto" v-if="isLoaded && items.length > 0">
 
       <b-table stacked="sm" stack hover :fields="FieldTableItems" :items="items" thead-class="thead-light">